chore: Add comments and cleanup unused translations script
