+++ package.json
@@ -170,7 +170,7 @@
     "translations": "npm-run-all translations:*",
     "translations:extract": "npx formatjs extract ./src/*.js --out-file ./build/messages/src/Messages.json",
     "translations:compile": "npx formatjs compile ./build/messages/src/Messages.json --out-file ./locales/en.json",
-    "translations:unused": "node ./scripts/unused.mjs"
+    "translations:unused": "node ./scripts/unusedTranslations.mjs"
   },
   "release": {
     "analyzeCommits": {
