--- scripts/unused.mjs
@@ -1,25 +1,0 @@
-import fs from 'fs';
-import { exec } from 'child_process';
-
-try {
-    const messagesFile = fs.readFileSync('./src/Messages.js', 'utf8')
-    let messagesFileLines = messagesFile.split('\n');
-
-    const startLine = messagesFileLines.indexOf('export default defineMessages({');
-    const endLine = messagesFileLines.indexOf('});');
-
-    messagesFileLines = messagesFileLines.slice(startLine, endLine + 1);
-    messagesFileLines[0] = "({"
-    messagesFileLines[messagesFileLines.length - 1] = "})"
-    messagesFileLines = messagesFileLines.join(' ');
-
-    const messages = eval(messagesFileLines)
-
-    Object.keys(messages).forEach(key => {
-        exec(`grep --include=*.js --exclude=*.test.js -rnw 'src' -e 'messages.${key}'`, error => {
-            error && console.log(`\x1b[31mUnused message key: ${key}\x1b[0m`);
-        });
-    });
-} catch (err) {
-    console.error(err)
-}
